Restructured pages and completed GitHub oauth
