+++ frontend/src/pages/manipulation/MenuManipulation.ts
