+++ frontend/src/pages/manipulation/VendorListManipulation.ts
